+++ test/integration/main_volume/controls/default.rb
@@ -1,20 +1,20 @@
-expected_hostname = input('expected_hostname')
+expected_os_family = input('expected_os_family')
 expected_os_name = input('expected_os_name')
 expected_os_version = input('expected_os_version')
-expected_disk_size = input('expected_disk_size')
-expected_memory = input('expected_memory')
-expected_vcpu = input('expected_vcpu')
+expected_disk_size = input('expected_disk_size').to_i
+expected_memory = input('expected_memory').to_i
+expected_vcpu = input('expected_vcpu').to_i
 
-describe bash('hostname') do
-  its('stdout') { should match(/#{expected_hostname}/) }
+describe os.family do
+  it { should eq expected_os_family }
 end
 
-describe bash('awk -F\'"\' \'/^NAME=/ {print $2}\' /etc/os-release |tr \'[:upper:]\' \'[:lower:]\'') do
-  its('stdout') { should match(/#{expected_os_name}/) }
+describe os.name do
+  it { should eq expected_os_name }
 end
 
-describe bash('awk -F\'"\' \'/^VERSION_ID=/ {print $2}\' /etc/os-release') do
-  its('stdout') { should match(/#{expected_os_version}/) }
+describe os.release do
+  it { should eq expected_os_version }
 end
 
 describe bash('df | grep root | awk \'{print $2}\' | tr -d "\n"') do