+++ CHANGELOG.md
@@ -1,5 +1,27 @@
 # Changelog
 
+## 0.0.2
+
+💥 Breaking changes:
+
+- Updating the provider version to 0.6.11
+
+🆕 New features:
+
+- Added Ubuntu 21.04 to the OS module
+
+- Autostart is now configurable for the domain resource
+
+🔧 Fixes:
+
+- Updated example and test suites
+
+([PR #2](https://github.com/srb3/terraform-libvirt-domain/pull/2))
+([PR #3](https://github.com/srb3/terraform-libvirt-domain/pull/3))
+([PR #4](https://github.com/srb3/terraform-libvirt-domain/pull/4))
+([PR #5](https://github.com/srb3/terraform-libvirt-domain/pull/5))
+
+
 ## 0.0.1
 
 💥 Breaking changes: